refactor: remove cjs builds
