+++ packages/documentation/src/components/layout.tsx
@@ -4,7 +4,7 @@ import * as React from 'react'
 import Helmet from 'react-helmet'
 import styled from 'styled-components'
 import HTML5Backend from 'react-dnd-html5-backend'
-import { isDebugMode } from 'react-dnd-examples-hooks/lib/esm/index'
+import { isDebugMode } from 'react-dnd-examples-hooks/lib/index'
 import { DndProvider } from 'react-dnd'
 import PageBody from './pagebody'
 import Sidebar from './sidebar'
