+++ packages/documentation/tsconfig.json
@@ -1,10 +1,11 @@
 {
-	"extends": "../../tsconfig.esm.base.json",
+	"extends": "../../tsconfig.json",
 	"compilerOptions": {
 		"outDir": "./lib",
 		"baseUrl": "./",
 		"allowSyntheticDefaultImports": true,
 		"esModuleInterop": true,
+		"declaration": false,
 		"allowJs": true
 	},
 	"include": ["./src/index.ts"]
