+++ packages/examples-hooks/package.json
@@ -3,8 +3,8 @@
 	"version": "7.6.0",
 	"private": true,
 	"description": "Drag and Drop for React",
-	"main": "lib/cjs/index.js",
-	"types": "lib/cjs/index.d.ts",
+	"main": "lib/index.js",
+	"types": "lib/index.d.ts",
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/react-dnd/react-dnd.git"
@@ -12,12 +12,11 @@
 	"license": "MIT",
 	"scripts": {
 		"clean": "rimraf lib",
-		"build:cjs": "tsc -b tsconfig.cjs.json",
-		"build:esm": "tsc -b tsconfig.esm.json",
+		"build:ts": "tsc -b tsconfig.json",
 		"build:docs": "tsc -b tsconfig.docs.json",
 		"build": "run-p build:*",
 		"test": "run-s clean build",
-		"start": "tsc -b tsconfig.cjs.json -w --preserveWatchOutput"
+		"start": "tsc -b tsconfig.json -w --preserveWatchOutput"
 	},
 	"dependencies": {
 		"@types/faker": "^4.1.5",
