--- packages/examples-hooks/tsconfig.cjs.json
@@ -1,10 +1,0 @@
-{
-	"extends": "../../tsconfig.cjs.base.json",
-	"compilerOptions": {
-		"allowSyntheticDefaultImports": true,
-		"esModuleInterop": true,
-		"outDir": "./lib/cjs",
-		"baseUrl": "./"
-	},
-	"include": ["./src/index.ts"]
-}
