+++ packages/react-dnd-decorators/src/DragLayer.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import checkDecoratorArguments from './utils/checkDecoratorArguments'
 import { DragDropManager, Unsubscribe } from 'dnd-core'
 import { DndContext, DndOptions } from 'react-dnd'
-import { isPlainObject } from 'react-dnd/lib/esm/internals'
+import { isPlainObject } from 'react-dnd/lib/internals'
 import { DragLayerCollector, DndComponentEnhancer } from './interfaces'
 import { isRefable } from './utils/isRefable'
