+++ packages/react-dnd-decorators/src/createSourceFactory.ts
@@ -3,7 +3,7 @@ declare var process: any
 import * as React from 'react'
 import { DragSource, DragDropMonitor } from 'dnd-core'
 import { DragSourceMonitor } from 'react-dnd'
-import { isPlainObject } from 'react-dnd/lib/esm/internals'
+import { isPlainObject } from 'react-dnd/lib/internals'
 import { DragSourceSpec } from './interfaces'
 import { getDecoratedComponent } from './utils/getDecoratedComponent'
