+++ packages/react-dnd-html5-backend/package.json
@@ -1,39 +1,38 @@
 {
-  "name": "react-dnd-html5-backend",
-  "version": "7.6.0",
-  "description": "HTML5 backend for React DnD",
-  "main": "lib/cjs/index.js",
-  "module": "lib/esm/index.js",
-  "types": "lib/cjs/index.d.ts",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/react-dnd/react-dnd.git"
-  },
-  "scripts": {
-    "clean": "rimraf lib dist",
-    "transpile:cjs": "tsc -b tsconfig.cjs.json",
-    "transpile:esm": "tsc -b tsconfig.esm.json",
-    "transpile": "run-p transpile:*",
-    "bundle:unmin": "webpack --mode development --output-filename=ReactDnDHTML5Backend.js",
-    "bundle:min": "webpack --mode production --output-filename=ReactDnDHTML5Backend.min.js",
-    "build": "run-p bundle:* transpile",
-    "test": "run-s clean build",
-    "start": "tsc -b tsconfig.esm.json -w --preserveWatchOutput"
-  },
-  "dependencies": {
-    "dnd-core": "^7.6.0"
-  },
-  "devDependencies": {
-    "@types/react": "^16.8.19",
-    "npm-run-all": "^4.1.5",
-    "react": "^16.8.6",
-    "react-dnd-test-backend": "^7.6.0",
-    "react-dom": "^16.8.6",
-    "rimraf": "^2.6.3",
-    "ts-loader": "^6.0.2",
-    "typescript": "^3.5.1",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2"
-  }
+	"name": "react-dnd-html5-backend",
+	"version": "7.6.0",
+	"description": "HTML5 backend for React DnD",
+	"main": "lib/index.js",
+	"module": "lib/index.js",
+	"types": "lib/index.d.ts",
+	"license": "MIT",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/react-dnd/react-dnd.git"
+	},
+	"scripts": {
+		"clean": "rimraf lib dist",
+		"transpile:esm": "tsc -b tsconfig.json",
+		"transpile": "run-p transpile:*",
+		"bundle:unmin": "webpack --mode development --output-filename=ReactDnDHTML5Backend.js",
+		"bundle:min": "webpack --mode production --output-filename=ReactDnDHTML5Backend.min.js",
+		"build": "run-p bundle:* transpile",
+		"test": "run-s clean build",
+		"start": "tsc -b tsconfig.json -w --preserveWatchOutput"
+	},
+	"dependencies": {
+		"dnd-core": "^7.6.0"
+	},
+	"devDependencies": {
+		"@types/react": "^16.8.19",
+		"npm-run-all": "^4.1.5",
+		"react": "^16.8.6",
+		"react-dnd-test-backend": "^7.6.0",
+		"react-dom": "^16.8.6",
+		"rimraf": "^2.6.3",
+		"ts-loader": "^6.0.2",
+		"typescript": "^3.5.1",
+		"webpack": "^4.32.2",
+		"webpack-cli": "^3.3.2"
+	}
 }
