+++ packages/react-dnd-test-backend/package.json
@@ -2,9 +2,9 @@
 	"name": "react-dnd-test-backend",
 	"version": "7.6.0",
 	"description": "A mock backend for testing React DnD apps",
-	"main": "lib/cjs/index.js",
-	"module": "lib/esm/index.js",
-	"types": "lib/cjs/index.d.ts",
+	"main": "lib/index.js",
+	"module": "lib/index.js",
+	"types": "lib/index.d.ts",
 	"license": "MIT",
 	"repository": {
 		"type": "git",
@@ -12,9 +12,7 @@
 	},
 	"scripts": {
 		"clean": "rimraf lib",
-		"build:cjs": "tsc -b tsconfig.cjs.json",
-		"build:esm": "tsc -b tsconfig.esm.json",
-		"build": "run-p build:*",
+		"build": "tsc -b tsconfig.json",
 		"test": "run-s clean build",
 		"prepublish": "npm run test"
 	},
