+++ packages/react-dnd-test-backend/tsconfig.json
@@ -1,7 +1,7 @@
 {
-	"extends": "../../tsconfig.esm.base.json",
+	"extends": "../../tsconfig.json",
 	"compilerOptions": {
-		"outDir": "./lib/esm",
+		"outDir": "./lib",
 		"baseUrl": "./src"
 	},
 	"include": ["src/index.ts"]
