+++ packages/react-dnd/package.json
@@ -2,9 +2,9 @@
 	"name": "react-dnd",
 	"version": "7.6.0",
 	"description": "Drag and Drop for React",
-	"main": "lib/cjs/index.js",
-	"module": "lib/esm/index.js",
-	"types": "lib/cjs/index.d.ts",
+	"main": "lib/index.js",
+	"module": "lib/index.js",
+	"types": "lib/index.d.ts",
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/react-dnd/react-dnd.git"
@@ -14,12 +14,11 @@
 		"clean": "rimraf lib dist",
 		"bundle:unmin": "webpack --mode development --output-filename=ReactDnD.js",
 		"bundle:min": "webpack --mode production --output-filename=ReactDnD.min.js",
-		"transpile:cjs": "tsc -b tsconfig.cjs.json",
-		"transpile:esm": "tsc -b tsconfig.esm.json",
+		"transpile:esm": "tsc -b tsconfig.json",
 		"transpile": "run-p transpile:*",
 		"build": "run-p bundle:* transpile",
 		"test": "run-s clean build",
-		"start": "tsc -b tsconfig.cjs.json -w --preserveWatchOutput"
+		"start": "tsc -b tsconfig.json -w --preserveWatchOutput"
 	},
 	"dependencies": {
 		"@types/hoist-non-react-statics": "^3.3.1",
