--- packages/react-dnd/tsconfig.cjs.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../../tsconfig.cjs.base.json",
-	"compilerOptions": {
-		"outDir": "./lib/cjs",
-		"baseUrl": "./"
-	},
-	"include": ["./src/index.ts", "./src/internals.ts"]
-}
