--- tsconfig.cjs.base.json
@@ -1,8 +1,0 @@
-{
-	"extends": "./tsconfig.esm.base.json",
-	"compilerOptions": {
-		"target": "es5",
-		"module": "commonjs",
-		"moduleResolution": "node"
-	}
-}
