--- tsconfig.esm.base.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"incremental": true,
-		"target": "es2015",
-		"module": "esnext",
-		"moduleResolution": "node",
-		"lib": ["esnext", "dom"],
-		"strict": true,
-		"experimentalDecorators": true,
-		"noUnusedLocals": true,
-		"jsx": "react",
-		"declaration": true
-	}
-}
